entry: move symlinks to a hashmap
To prepare for other sort options which require the statx field (ie time
sorting), move the symlink data off of Entry to reduce it's size. This
improves sorting speed due to the Entry struct being smaller in memory.
Symlinks are generally rare so using a hashmap should be ok. This was
tested locally in /usr/bin which has a decent amount of symlinks and no
noticeable performance hit was seen as a result of this approach.